\name{MOGC-package}
\alias{MOGC-package}
\alias{MOGC}
\docType{package}
\title{
\packageTitle{MOGC}
}
\description{
\packageDescription{MOGC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MOGC}
\packageIndices{MOGC}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MOGC}

Maintainer: \packageMaintainer{MOGC}
}

