% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOGC_callC.R
\name{MOGC}
\alias{MOGC}
\title{Bayesian Gaussian mixture model incorporating Multi-layer Overlapping Group structure (MOGC)}
\usage{
MOGC(Y, U1, U2, K, center = FALSE, burnInIter = 1000,
  keepIter = 3000, maxIter = 10000, print_int = 1000,
  debug = FALSE, init_mu = NULL, init_z = NULL, fix_z = FALSE,
  fix_mu = FALSE, adj_ls = TRUE, seed = 123,
  BernoulliWeighted = TRUE, MH_ind = 0, pi_g_prop_n = 10,
  pi_j_prop_n = 10, s2 = NULL, types = NULL,
  report_dup_mean = FALSE)
}
\arguments{
\item{Y}{n by p feature matrix, where n is the sample size and p is feature size}

\item{U1}{p by m1 matrix denoting feature group membership}

\item{U2}{m1 by m2 matrix denoting level-2 group membership matrix of level-1 groups}

\item{K}{number of clusters}

\item{center}{TRUE/FALSE, if each column of Y should be centered}

\item{burnInIter}{number of iterations as burn-in period}

\item{keepIter}{number of iterations to be stored}

\item{maxIter}{maximum number of iterations}

\item{print_int}{print progress message every print_int iterations}

\item{debug}{TRUE/FALSE, indicating debug mode}

\item{init_mu}{K by p matrix, indicate the initial mu}

\item{init_z}{vector of length n, indicating cluster label for initial}

\item{fix_z}{TRUE/FALSE, if TRUE, do not update cluster label (for debugging)}

\item{fix_mu}{TRUE/FALSE, if TRUE, do not update cluster mean (for debugging)}

\item{adj_ls}{TRUE/FALSE, if TRUE, adjust for label switching using package label.switching}

\item{seed}{integer as random seed}

\item{BernoulliWeighted}{TRUE/FALSE, if true, adding R_j/D_g for fairness (see paper)}

\item{MH_ind}{TRUE/FALSE, if true, MH is used to update pi_gk. If overlapping group exists, MH_ind must be specified as TRUE}

\item{pi_g_prop_n}{integer, specifying n_0 in the proposal distribution of pi_lk for MH updating}

\item{pi_j_prop_n}{integer, specifying n_0 in the proposal distribution of pi_glk for MH updating}

\item{s2}{numeric number, s2 will be fixed and not be updated}

\item{types}{a vector of length p, indicating the types of each feature. Type-specific s2 will be estimated.}

\item{report_dup_mean}{TRUE/FALSE, if true, the v_jgk will be returned}
}
\value{
A list of MCMC output
}
\description{
This function applies MOGC clustering model to a feature matrix
}
