% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuMOGC.R
\name{simuMOGC}
\alias{simuMOGC}
\title{Simulate data with MOGC structure}
\usage{
simuMOGC(seed = 15213, K = 3, numSamplesPerK, U2, l_index,
  numFeatInLevel1Group = 3, strongGroupSignal,
  weakGroupSignalRatio = 0.5, strongGroupPerc, weakGroupPerc,
  noiseMean = 0, sample_sigma, sigma_noise, percConfounder = 0, rho1,
  rho2, random_mu = FALSE)
}
\arguments{
\item{random_mu}{}
}
\description{
Simulate data with MOGC structure
}
