% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuSOGC.R
\name{simuSOGC}
\alias{simuSOGC}
\title{Simulate data with SOGC structure}
\usage{
simuSOGC(seed = 15213, K = 3, numSamplesPerK, U1, g_index,
  strongSignal, weakSignalRatio = 0.5, strongSignalPerc, weakSignalPerc,
  noiseMean = 0, sample_sigma, sigma_noise, percConfounder = 0,
  rho = 0, random_mu = FALSE)
}
\arguments{
\item{random_mu}{}
}
\description{
Simulate data with SOGC structure
}
